+++ src/FluxFlow.Components.Routing/RoutingComponentPorts.cs
@@ -3,6 +3,8 @@ namespace FluxFlow.Components.Routing;
 public static class RoutingComponentPorts
 {
     public const string Input = "Input";
+    public const string Request = "Request";
+    public const string Response = "Response";
     public const string Left = "Left";
     public const string Right = "Right";
     public const string Output = "Output";